[Refactor] Simplify and reuse some code from model auto tracing.
